Added new module: library.lua
This module is a general purpose module that can be used to fetch a library from any parent mod.

Can be useful when several mods in a modpack use the same library.

If someone would want a mod as a standalone mod, they would have to copy any required libraries from the parent mod into the standalone mod.
In addition to this, if a library does not automatically init/load, that code would have to be copied into the parent mod as well.

Potential for improvement could be to have the possibility for mods to have a small file for init/load instructions for its libraries, in the case where someone would want to rearrange the folder structure of mods.