Add SetFire, SetForest and SetForestFire
`SetForest` is the same as setting `TERRAIN_FOREST` (6) with `Board:SetTerrain`, except it will correctly update the tile's fire type so the correct graphics will be displayed if the tile is on fire.

`SetForestFire` changes the tile's terrain to `TERRAIN_ROAD` (0) and applies the fire type `FIRE_TYPE_FOREST_FIRE` (2) in order to make the tile display the correct forest fire graphics.

`SetFire` is the same as the mod loader's `Board:SetFire` (renamed `Board:SetFireVanilla`), except it will update the fire type in order to make the tile display the correct fire graphics.